- Fix some format specifier mismatches with parameter types in sprintf... - Make CSIS3820 64 bit clean.

+++ sbs/vmemodules/CADC2530.cpp
@@ -388,7 +388,8 @@ uint16_t CADC2530::volt2uld(double volt) {
 const std::string& CADC2530::toString() {
   static string namstr;
   static char buffer[256];
-  sprintf(buffer,"ADC2530 in crate %d at base 0x%lx with memory offset 0x%lx",my_nCrate,my_nBase,my_nMemOffset);
+  sprintf(buffer,"ADC2530 in crate %d at base 0x%lx with memory offset 0x%lx",
+	  my_nCrate,(unsigned  long)my_nBase,(unsigned long)my_nMemOffset);
   namstr = buffer;
   return namstr;
 }
@@ -632,7 +633,9 @@ void CADC2530::mapModule() {
    // Donot map the register memory more than once.
    if (my_pModule != NULL) {
      char buffer[256];
-     sprintf(buffer, "CADC2530::mapModule(): The Card in crate %d at base 0x%lx already has a register memory map.\n",my_nCrate,my_nBase);
+     sprintf(buffer, 
+	     "CADC2530::mapModule(): The Card in crate %d at base 0x%lx already has a register memory map.\n",
+	     my_nCrate,(unsigned long)my_nBase);
      throw string(buffer);
    }
 
@@ -659,7 +662,7 @@ void CADC2530::mapModule() {
      my_pModule = NULL;
      char buffer[256];
      sprintf(buffer, "CADC2530::mapModule(): Card in crate %d at base 0x%lx is not a Hytec 2530 ADC or is missing type=%d id=0x%0x\n",
-	     my_nCrate,my_nBase,my_nCardType,
+	     my_nCrate, static_cast<unsigned long>(my_nBase), my_nCardType,
 	     static_cast<unsigned int>(my_nCardId));
      throw string(buffer);
    }
@@ -685,14 +688,17 @@ void CADC2530::mapMemory() {
    // We must have a register memory map first.
    if (my_pModule == NULL) {
      char buffer[256];
-     sprintf(buffer, "CADC2530::mapMemory(): Cannot map list/histogram memory without first mapping the ADC2530 registers for card in crate %d at base 0x%lx.\n",my_nCrate,my_nBase);
+     sprintf(buffer, 
+	     "CADC2530::mapMemory(): Cannot map list/histogram memory without first mapping the ADC2530 registers for card in crate %d at base 0x%lx.\n",
+	     my_nCrate, (unsigned long)my_nBase);
      throw string(buffer);
    } 
 
    // Donot map the list/histogram memory more than once.
    if (my_pMemory != NULL) {
      char buffer[256];
-     sprintf(buffer, "CADC2530::mapMemory(): The Card in crate %d at base 0x%lx already has a list/histogram memory map.\n",my_nCrate,my_nBase);
+     sprintf(buffer, "CADC2530::mapMemory(): The Card in crate %d at base 0x%lx already has a list/histogram memory map.\n",
+	     my_nCrate, (unsigned long)my_nBase);
      throw string(buffer);
    }
 
@@ -711,7 +717,8 @@ void CADC2530::mapMemory() {
    // we should read back the setting we calculated.
    if (nmemoset != my_nMemOffset) {
      char buffer[256];
-     sprintf(buffer, "CADC2530::mapMemory(): Error setting list/histogram memory offset for ADC2530 card in crate %d at base 0x%lx (written=0x%lx, read=0x%lx)\n",my_nCrate,my_nBase,nmemoset,my_nMemOffset);
+     sprintf(buffer, "CADC2530::mapMemory(): Error setting list/histogram memory offset for ADC2530 card in crate %d at base 0x%lx (written=0x%lx, read=0x%lx)\n",my_nCrate,
+	     (unsigned long)my_nBase, (unsigned long)nmemoset, (unsigned long)my_nMemOffset);
      throw string(buffer);
    } 
 
@@ -758,7 +765,8 @@ void CADC2530::clearMemory() {
   // We must have a memory map first.
   if (my_pMemory == NULL) {
     char buffer[256];
-    sprintf(buffer, "CADC2530::clearMemory(): Cannot clear list/histogram memory without first having a memory map for card in crate %d at base 0x%lx.\n",my_nCrate,my_nBase);
+    sprintf(buffer, "CADC2530::clearMemory(): Cannot clear list/histogram memory without first having a memory map for card in crate %d at base 0x%lx.\n",my_nCrate,
+	    (unsigned long)my_nBase);
     throw string(buffer);
   } 
 
@@ -785,7 +793,8 @@ void CADC2530::clearHistogramMemory() {
   // We must have a memory map first.
   if (my_pMemory == NULL) {
     char buffer[256];
-    sprintf(buffer, "CADC2530::clearHistogramMemory(): Cannot clear histogram memory without first having a memory map for card in crate %d at base 0x%lx.\n",my_nCrate,my_nBase);
+    sprintf(buffer, "CADC2530::clearHistogramMemory(): Cannot clear histogram memory without first having a memory map for card in crate %d at base 0x%lx.\n",
+	    my_nCrate, (unsigned long)my_nBase);
     throw string(buffer);
   } 
 
@@ -812,7 +821,8 @@ void CADC2530::resetCard() {
   // We must have a memory map first.
   if (my_pModule == NULL) {
     char buffer[256];
-    sprintf(buffer, "CADC2530::resetCard(): Cannot reset this card without first having a memory map for card in crate %d at base 0x%lx.\n",my_nCrate,my_nBase);
+    sprintf(buffer, "CADC2530::resetCard(): Cannot reset this card without first having a memory map for card in crate %d at base 0x%lx.\n",
+	    my_nCrate, (unsigned long)my_nBase);
     throw string(buffer);
   }
